requirements.txt
streamlit==1.23.1
requests==2.31.0
beautifulsoup4==4.12.2
faiss-cpu==1.7.3
sentence-transformers==2.2.2
langchain==0.0.267
